requirements.txt
anthropic>=0.18.0
colorama>=0.4.6
python-dotenv>=1.0.0
instructor>=0.6.0
pydantic>=2.0.0
prompt_toolkit>=3.0.43 

